Non-mentor review of third task
