+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.BasicApplication"
         tools:targetApi="31">
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".presentation.ui.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
