--- app/src/main/java/com/shpp/ahrokholska/basicapplication/data/Contact.kt
@@ -1,6 +1,0 @@
-package com.shpp.ahrokholska.basicapplication.data
-
-import java.util.UUID
-
-data class Contact(val picture:String, val name: String, val career: String,
-                   val id:Long = UUID.randomUUID().mostSignificantBits)
