+++ app/src/main/java/com/shpp/ahrokholska/basicapplication/data/repository/UserRepositoryImpl.kt
@@ -1,0 +1,33 @@
+package com.shpp.ahrokholska.basicapplication.data.repository
+
+import android.content.Context
+import androidx.datastore.preferences.core.edit
+import com.shpp.ahrokholska.basicapplication.domain.repository.UserRepository
+import com.shpp.ahrokholska.basicapplication.presentation.utils.Constants
+import com.shpp.ahrokholska.basicapplication.data.utils.ext.dataStore
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
+
+// todo look at me
+class UserRepositoryImpl(private val context: Context) : UserRepository {
+    override val userName: Flow<String> = context.dataStore.data.map { preferences ->
+        preferences[Constants.STORED_USER_NAME_KEY].orEmpty()
+    }
+
+    override val isAutoLoginEnabled: Flow<Boolean> = context.dataStore.data.map { preferences ->
+        preferences[Constants.IS_AUTO_LOGIN_ENABLED_KEY] ?: false
+    }
+
+    override suspend fun saveUsername(value: String) {
+        context.dataStore.edit { settings ->
+            settings[Constants.STORED_USER_NAME_KEY] = value
+        }
+    }
+
+    // TODO is it violation of SOLID or not?
+    override suspend fun saveAutoLoginState(isAutoLoginEnabled: Boolean) {
+        context.dataStore.edit { settings ->
+            settings[Constants.IS_AUTO_LOGIN_ENABLED_KEY] = isAutoLoginEnabled
+        }
+    }
+}
