+++ app/src/main/java/com/shpp/ahrokholska/basicapplication/domain/repository/ContactsRepository.kt
@@ -1,0 +1,13 @@
+package com.shpp.ahrokholska.basicapplication.domain.repository
+
+import com.shpp.ahrokholska.basicapplication.domain.model.Contact
+import kotlinx.coroutines.flow.StateFlow
+
+interface ContactsRepository {
+    val contacts: StateFlow<List<Contact>>
+
+    suspend fun getContactWithId(id: Long): Contact
+    suspend fun addContact(name: String, career: String)
+    suspend fun removeWithId(id: Long)
+    suspend fun insertAt(contact: Contact, position: Int)
+}
