+++ app/src/main/java/com/shpp/ahrokholska/basicapplication/presentation/ui/addContact/AddContactViewModel.kt
@@ -1,0 +1,18 @@
+package com.shpp.ahrokholska.basicapplication.presentation.ui.addContact
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.shpp.ahrokholska.basicapplication.data.repository.HardcodedContactsRepositoryImpl
+import com.shpp.ahrokholska.basicapplication.domain.repository.ContactsRepository
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+
+class AddContactViewModel : ViewModel() {
+    private val contactsRepository: ContactsRepository = HardcodedContactsRepositoryImpl()
+
+    fun addContact(name: String, career: String) {
+        viewModelScope.launch(Dispatchers.IO) {
+            contactsRepository.addContact(name, career)
+        }
+    }
+}
