+++ app/src/main/java/com/shpp/ahrokholska/basicapplication/presentation/ui/contacts/adapter/diffCallback/ContactsDiffCallback.kt
@@ -1,0 +1,12 @@
+package com.shpp.ahrokholska.basicapplication.presentation.ui.contacts.adapter.diffCallback
+
+import androidx.recyclerview.widget.DiffUtil
+import com.shpp.ahrokholska.basicapplication.domain.model.Contact
+
+class ContactsDiffCallback : DiffUtil.ItemCallback<Contact>() {
+    override fun areItemsTheSame(oldItem: Contact, newItem: Contact): Boolean =
+        oldItem.id == newItem.id
+
+    override fun areContentsTheSame(oldItem: Contact, newItem: Contact): Boolean =
+        oldItem == newItem
+}
