+++ app/src/main/java/com/shpp/ahrokholska/basicapplication/presentation/utils/VerticalSpaceItemDecoration.kt
@@ -1,0 +1,18 @@
+package com.shpp.ahrokholska.basicapplication.presentation.utils
+
+import android.graphics.Rect
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+
+class VerticalSpaceItemDecoration(private val space: Int) : RecyclerView.ItemDecoration() {
+    override fun getItemOffsets(
+        outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State
+    ) {
+        outRect.apply {
+            bottom = space
+            if (parent.getChildAdapterPosition(view) == 0) {
+                top = space
+            }
+        }
+    }
+}
