--- app/src/main/java/com/shpp/ahrokholska/basicapplication/ui/MainActivity.kt
@@ -1,14 +1,0 @@
-package com.shpp.ahrokholska.basicapplication.ui
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.shpp.ahrokholska.basicapplication.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-    private val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(binding.root)
-    }
-}
